Update sorting of `po.compile()` to accept a sort function.
The `sortByMsgId` option is removed and replaced with `sort`.

If `sort` is `true`, the `sortByMsgId` algorithm is used for sorting the output.
This is based on the `--sort-output` argument of the gnu xgettext cli.

If `sort` is a function, that function is used as the comparator function
for sorting the output.

By default, `sort` is `false` (consistent with the previous `sortByMsgId` and gnu gettext).