fix: foldLine escape character logic
The escape character logic was not working correctly.

The logic of how a line is folded with a final escape character was
flawed. Using a loop to consume forward slashes would not allow for an
escaped forward slash to be folded correctly, it would append an extra
character to the line.

- Use if statements to check for escape characters. Only one extra
  character should be appended to the line, and it should be considered
  escaped.
- Add new test cases to test the escape character logic.